FIXME comments in LIGO_simulation.mlx and analyzing_datasignals.mlx need to be addressed.
The paths to SIGNALS and NOISE directories in SNRcalcMod2.m needed to be fixed.

SNRcalcMod2.m is giving me an estimated SNR plot of NaN (not a number).
In your code at line 54, the value of normSigSqrd is infinity.
I suspect this is because your input PSD anomalous values of zero at the beginning.

In data_glrt_calc.m, you have not completed the last step in Task 4: estimating the significance.